chore: use one kms key for storage and db, and go back to one stack
